Struct delete_item::Client

pub(crate) struct Client<C = DynConnector, M = AwsMiddleware, R = Standard> {
    handle: Arc<Handle<C, M, R>>,
}

An ergonomic service client for DynamoDB_20120810.

This client allows ergonomic access to a DynamoDB_20120810-shaped service. Each method corresponds to an endpoint defined in the service’s Smithy model, and the request and response shapes are auto-generated from that same model.

Using a Client

Once you have a client set up, you can access the service’s endpoints by calling the appropriate method on Client. Each such method returns a request builder for that endpoint, with methods for setting the various fields of the request. Once your request is complete, use the send method to send the request. send returns a future, which you then have to .await to get the service’s response.

Fields
handle: Arc<Handle<C, M, R>>
Implementations

impl<C, M, R> Client<C, M, R>

pub fn with_config(client: Client<C, M, R>, conf: Config) -> Client<C, M, R>

Creates a client with the given service configuration.

pub fn conf(&self) -> &Config

Returns the client’s configuration.

impl<C, M, R> Client<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub fn batch_execute_statement(&self) -> BatchExecuteStatement<C, M, R>

Constructs a fluent builder for the BatchExecuteStatement operation.

See BatchExecuteStatement for more information about the operation and its arguments.

pub fn batch_get_item(&self) -> BatchGetItem<C, M, R>

Constructs a fluent builder for the BatchGetItem operation.

See BatchGetItem for more information about the operation and its arguments.

pub fn batch_write_item(&self) -> BatchWriteItem<C, M, R>

Constructs a fluent builder for the BatchWriteItem operation.

See BatchWriteItem for more information about the operation and its arguments.

pub fn create_backup(&self) -> CreateBackup<C, M, R>

Constructs a fluent builder for the CreateBackup operation.

See CreateBackup for more information about the operation and its arguments.

pub fn create_global_table(&self) -> CreateGlobalTable<C, M, R>

Constructs a fluent builder for the CreateGlobalTable operation.

See CreateGlobalTable for more information about the operation and its arguments.

pub fn create_table(&self) -> CreateTable<C, M, R>

Constructs a fluent builder for the CreateTable operation.

See CreateTable for more information about the operation and its arguments.

pub fn delete_backup(&self) -> DeleteBackup<C, M, R>

Constructs a fluent builder for the DeleteBackup operation.

See DeleteBackup for more information about the operation and its arguments.

pub fn delete_item(&self) -> DeleteItem<C, M, R>

Constructs a fluent builder for the DeleteItem operation.

See DeleteItem for more information about the operation and its arguments.

pub fn delete_table(&self) -> DeleteTable<C, M, R>

Constructs a fluent builder for the DeleteTable operation.

See DeleteTable for more information about the operation and its arguments.

pub fn describe_backup(&self) -> DescribeBackup<C, M, R>

Constructs a fluent builder for the DescribeBackup operation.

See DescribeBackup for more information about the operation and its arguments.

pub fn describe_continuous_backups(&self) -> DescribeContinuousBackups<C, M, R>

Constructs a fluent builder for the DescribeContinuousBackups operation.

See DescribeContinuousBackups for more information about the operation and its arguments.

pub fn describe_contributor_insights(
&self
) -> DescribeContributorInsights<C, M, R>

Constructs a fluent builder for the DescribeContributorInsights operation.

See DescribeContributorInsights for more information about the operation and its arguments.

pub fn describe_endpoints(&self) -> DescribeEndpoints<C, M, R>

Constructs a fluent builder for the DescribeEndpoints operation.

See DescribeEndpoints for more information about the operation and its arguments.

pub fn describe_export(&self) -> DescribeExport<C, M, R>

Constructs a fluent builder for the DescribeExport operation.

See DescribeExport for more information about the operation and its arguments.

pub fn describe_global_table(&self) -> DescribeGlobalTable<C, M, R>

Constructs a fluent builder for the DescribeGlobalTable operation.

See DescribeGlobalTable for more information about the operation and its arguments.

pub fn describe_global_table_settings(
&self
) -> DescribeGlobalTableSettings<C, M, R>

Constructs a fluent builder for the DescribeGlobalTableSettings operation.

See DescribeGlobalTableSettings for more information about the operation and its arguments.

pub fn describe_kinesis_streaming_destination(
&self
) -> DescribeKinesisStreamingDestination<C, M, R>

Constructs a fluent builder for the DescribeKinesisStreamingDestination operation.

See DescribeKinesisStreamingDestination for more information about the operation and its arguments.

pub fn describe_limits(&self) -> DescribeLimits<C, M, R>

Constructs a fluent builder for the DescribeLimits operation.

See DescribeLimits for more information about the operation and its arguments.

pub fn describe_table(&self) -> DescribeTable<C, M, R>

Constructs a fluent builder for the DescribeTable operation.

See DescribeTable for more information about the operation and its arguments.

pub fn describe_table_replica_auto_scaling(
&self
) -> DescribeTableReplicaAutoScaling<C, M, R>

Constructs a fluent builder for the DescribeTableReplicaAutoScaling operation.

See DescribeTableReplicaAutoScaling for more information about the operation and its arguments.

pub fn describe_time_to_live(&self) -> DescribeTimeToLive<C, M, R>

Constructs a fluent builder for the DescribeTimeToLive operation.

See DescribeTimeToLive for more information about the operation and its arguments.

pub fn disable_kinesis_streaming_destination(
&self
) -> DisableKinesisStreamingDestination<C, M, R>

Constructs a fluent builder for the DisableKinesisStreamingDestination operation.

See DisableKinesisStreamingDestination for more information about the operation and its arguments.

pub fn enable_kinesis_streaming_destination(
&self
) -> EnableKinesisStreamingDestination<C, M, R>

Constructs a fluent builder for the EnableKinesisStreamingDestination operation.

See EnableKinesisStreamingDestination for more information about the operation and its arguments.

pub fn execute_statement(&self) -> ExecuteStatement<C, M, R>

Constructs a fluent builder for the ExecuteStatement operation.

See ExecuteStatement for more information about the operation and its arguments.

pub fn execute_transaction(&self) -> ExecuteTransaction<C, M, R>

Constructs a fluent builder for the ExecuteTransaction operation.

See ExecuteTransaction for more information about the operation and its arguments.

pub fn export_table_to_point_in_time(&self) -> ExportTableToPointInTime<C, M, R>

Constructs a fluent builder for the ExportTableToPointInTime operation.

See ExportTableToPointInTime for more information about the operation and its arguments.

pub fn get_item(&self) -> GetItem<C, M, R>

Constructs a fluent builder for the GetItem operation.

See GetItem for more information about the operation and its arguments.

pub fn list_backups(&self) -> ListBackups<C, M, R>

Constructs a fluent builder for the ListBackups operation.

See ListBackups for more information about the operation and its arguments.

pub fn list_contributor_insights(&self) -> ListContributorInsights<C, M, R>

Constructs a fluent builder for the ListContributorInsights operation.

See ListContributorInsights for more information about the operation and its arguments.

pub fn list_exports(&self) -> ListExports<C, M, R>

Constructs a fluent builder for the ListExports operation.

See ListExports for more information about the operation and its arguments.

pub fn list_global_tables(&self) -> ListGlobalTables<C, M, R>

Constructs a fluent builder for the ListGlobalTables operation.

See ListGlobalTables for more information about the operation and its arguments.

pub fn list_tables(&self) -> ListTables<C, M, R>

Constructs a fluent builder for the ListTables operation.

See ListTables for more information about the operation and its arguments.

pub fn list_tags_of_resource(&self) -> ListTagsOfResource<C, M, R>

Constructs a fluent builder for the ListTagsOfResource operation.

See ListTagsOfResource for more information about the operation and its arguments.

pub fn put_item(&self) -> PutItem<C, M, R>

Constructs a fluent builder for the PutItem operation.

See PutItem for more information about the operation and its arguments.

pub fn query(&self) -> Query<C, M, R>

Constructs a fluent builder for the Query operation.

See Query for more information about the operation and its arguments.

pub fn restore_table_from_backup(&self) -> RestoreTableFromBackup<C, M, R>

Constructs a fluent builder for the RestoreTableFromBackup operation.

See RestoreTableFromBackup for more information about the operation and its arguments.

pub fn restore_table_to_point_in_time(
&self
) -> RestoreTableToPointInTime<C, M, R>

Constructs a fluent builder for the RestoreTableToPointInTime operation.

See RestoreTableToPointInTime for more information about the operation and its arguments.

pub fn scan(&self) -> Scan<C, M, R>

Constructs a fluent builder for the Scan operation.

See Scan for more information about the operation and its arguments.

pub fn tag_resource(&self) -> TagResource<C, M, R>

Constructs a fluent builder for the TagResource operation.

See TagResource for more information about the operation and its arguments.

pub fn transact_get_items(&self) -> TransactGetItems<C, M, R>

Constructs a fluent builder for the TransactGetItems operation.

See TransactGetItems for more information about the operation and its arguments.

pub fn transact_write_items(&self) -> TransactWriteItems<C, M, R>

Constructs a fluent builder for the TransactWriteItems operation.

See TransactWriteItems for more information about the operation and its arguments.

pub fn untag_resource(&self) -> UntagResource<C, M, R>

Constructs a fluent builder for the UntagResource operation.

See UntagResource for more information about the operation and its arguments.

pub fn update_continuous_backups(&self) -> UpdateContinuousBackups<C, M, R>

Constructs a fluent builder for the UpdateContinuousBackups operation.

See UpdateContinuousBackups for more information about the operation and its arguments.

pub fn update_contributor_insights(&self) -> UpdateContributorInsights<C, M, R>

Constructs a fluent builder for the UpdateContributorInsights operation.

See UpdateContributorInsights for more information about the operation and its arguments.

pub fn update_global_table(&self) -> UpdateGlobalTable<C, M, R>

Constructs a fluent builder for the UpdateGlobalTable operation.

See UpdateGlobalTable for more information about the operation and its arguments.

pub fn update_global_table_settings(&self) -> UpdateGlobalTableSettings<C, M, R>

Constructs a fluent builder for the UpdateGlobalTableSettings operation.

See UpdateGlobalTableSettings for more information about the operation and its arguments.

pub fn update_item(&self) -> UpdateItem<C, M, R>

Constructs a fluent builder for the UpdateItem operation.

See UpdateItem for more information about the operation and its arguments.

pub fn update_table(&self) -> UpdateTable<C, M, R>

Constructs a fluent builder for the UpdateTable operation.

See UpdateTable for more information about the operation and its arguments.

pub fn update_table_replica_auto_scaling(
&self
) -> UpdateTableReplicaAutoScaling<C, M, R>

Constructs a fluent builder for the UpdateTableReplicaAutoScaling operation.

See UpdateTableReplicaAutoScaling for more information about the operation and its arguments.

pub fn update_time_to_live(&self) -> UpdateTimeToLive<C, M, R>

Constructs a fluent builder for the UpdateTimeToLive operation.

See UpdateTimeToLive for more information about the operation and its arguments.

impl<C> Client<C, AwsMiddleware, Standard>

pub fn from_conf_conn(
conf: Config,
conn: C
) -> Client<C, AwsMiddleware, Standard>

Creates a client with the given service config and connector override.

impl Client<DynConnector, AwsMiddleware, Standard>

pub fn new(config: &Config) -> Client<DynConnector, AwsMiddleware, Standard>

Creates a new client from a shared config.

pub fn from_conf(conf: Config) -> Client<DynConnector, AwsMiddleware, Standard>

Creates a new client from the service Config.